+++ bigmler/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import absolute_import
 
-__version__ = '3.5.1'
+__version__ = '3.5.2'
 
 if __name__ == '__main__':
     import sys